+++ src/main/java/fi/dwo/server/PersistentDataManagers/core/StudentScoDataManager.java
@@ -70,7 +70,7 @@ public class StudentScoDataManager {
         } catch (Exception e) {
             String msg = e.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Integer id = ssd.getStudentSco();
+                Long id = ssd.getStudentSco();
                 if (findEntity(id) == null) {
                     LOG.log(Level.FINE, "The PersistentStudentScoData with " + id + " no longer exists.", e);
                     throw new PersistenceException(e);
@@ -89,7 +89,7 @@ public class StudentScoDataManager {
      *
      * @param id
      */
-    public static void destroy(Integer id) throws PersistenceException {
+    public static void destroy(Long id) throws PersistenceException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -135,7 +135,7 @@ public class StudentScoDataManager {
         }
     }
 
-    public static PersistentStudentScoData findEntity(Integer id) {
+    public static PersistentStudentScoData findEntity(Long id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(PersistentStudentScoData.class, id);
